refactor: move random picture to directive
DOM manipulations are better off there
