+++ client/app/components/_profilePicture/profilePicture.html
@@ -1,3 +1,4 @@
 <a ui-sref="userDetail({id:$ctrl.userId})">
+  <random-picture seed="$ctrl.userId" size="$ctrl.size"></random-picture>
   <md-tooltip>{{ $ctrl.name }}</md-tooltip>
 </a>
