+++ client/app/components/_profilePicture/profilePicture.js
@@ -1,6 +1,7 @@
 import angular from "angular";
 import uiRouter from "angular-ui-router";
 import profilePictureComponent from "./profilePicture.component";
+import randomPictureDirective from "./randomPicture.directive";
 import userService from "../../services/user/user";
 
 let profilePictureModule = angular.module("profilePicture", [
@@ -9,6 +10,8 @@ let profilePictureModule = angular.module("profilePicture", [
 ])
 
 .component("profilePicture", profilePictureComponent)
+
+.directive("randomPicture", randomPictureDirective)
 
 .name;
